+++ api-extension/acl-impl/partner-cluster/partner-cluster-multi/src/main/java/cn/tongdun/kunpeng/api/infrastructure/persistence/mybatis/mappers/kunpeng/PartnerClusterDAO.java
@@ -20,5 +20,17 @@ public interface PartnerClusterDAO {
      */
     List<PartnerClusterDO> selectAllAvailable();
 
+    /**
+     * 根据uuid查询数据
+     * @param uuid
+     * @return
+     */
+    PartnerClusterDO selectByUuid(String uuid);
 
+    /**
+     * 根据合作方名称查询合作方信息
+     * @param partnerCode
+     * @return
+     */
+    PartnerClusterDO selectByPartnerCode(String partnerCode);
 }